if packages check before deleteAll()

+++ api/modules/v1/controllers/OrderController.php
@@ -607,33 +607,33 @@ class OrderController extends ControllerEx
             }
 
             // Packages
-            PackageEx::deleteAll(['order_id' => $order->id]);
-            foreach ($orderForm->packages as $formPackage) {
-                \Yii::debug($formPackage);
-                $package = new PackageEx();
-                $package->setAttribute('length', $formPackage['length']);
-                $package->setAttribute('width', $formPackage['width']);
-                $package->setAttribute('height', $formPackage['height']);
-                $package->setAttribute('tracking', $formPackage['tracking']);
-                $package->setAttribute('order_id', $order->id);
-                $package->save();
-                foreach ($formPackage['package_items'] as $package_item) {
-                    $packageItem = new PackageItem();
-                    $packageItem->setAttribute('quantity', $package_item['quantity']);
-                    $packageItem->setAttribute('sku', $package_item['sku']);
-                    $packageItem->setAttribute('name', $package_item['name']);
-                    $packageItem->setAttribute('package_id', $package->id);
-                    $packageItem->setAttribute('order_id', $order->id);
-                    $packageItem->save();
-                    foreach ($package_item['lot_info'] as $lot_info) {
 
-
-                        $lotInfo = new PackageItemLotInfo();
-                        $lotInfo->setAttribute('quantity', $lot_info['quantity']);
-                        $lotInfo->setAttribute('lot_number', $lot_info['lot_number']);
-                        $lotInfo->setAttribute('serial_number', $lot_info['serial_number']);
-                        $lotInfo->setAttribute('package_items_id', $packageItem->id);
-                        $lotInfo->save();
+            if (!empty($orderForm->packages)) {
+                PackageEx::deleteAll(['order_id' => $order->id]);
+                foreach ($orderForm->packages as $formPackage) {
+                    $package = new PackageEx();
+                    $package->setAttribute('length', $formPackage['length']);
+                    $package->setAttribute('width', $formPackage['width']);
+                    $package->setAttribute('height', $formPackage['height']);
+                    $package->setAttribute('tracking', $formPackage['tracking']);
+                    $package->setAttribute('order_id', $order->id);
+                    $package->save();
+                    foreach ($formPackage['package_items'] as $package_item) {
+                        $packageItem = new PackageItem();
+                        $packageItem->setAttribute('quantity', $package_item['quantity']);
+                        $packageItem->setAttribute('sku', $package_item['sku']);
+                        $packageItem->setAttribute('name', $package_item['name']);
+                        $packageItem->setAttribute('package_id', $package->id);
+                        $packageItem->setAttribute('order_id', $order->id);
+                        $packageItem->save();
+                        foreach ($package_item['lot_info'] as $lot_info) {
+                            $lotInfo = new PackageItemLotInfo();
+                            $lotInfo->setAttribute('quantity', $lot_info['quantity']);
+                            $lotInfo->setAttribute('lot_number', $lot_info['lot_number']);
+                            $lotInfo->setAttribute('serial_number', $lot_info['serial_number']);
+                            $lotInfo->setAttribute('package_items_id', $packageItem->id);
+                            $lotInfo->save();
+                        }
                     }
                 }
             }
